# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Inferno cloud storage API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ItemType = typeof ItemType[keyof typeof ItemType];


export const ItemType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface Item {
  id: number;
  name: string;
  type: ItemType;
  /** @nullable */
  parentId: number | null;
  /** @nullable */
  objectPath?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type ItemInputType = typeof ItemInputType[keyof typeof ItemInputType];


export const ItemInputType = {
  file: 'file',
  folder: 'folder',
} as const;

export interface ItemInput {
  /** @minLength 1 */
  name: string;
  type: ItemInputType;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  objectPath?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
}

export interface ItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  parentId?: number | null;
}

export type PathSegmentType = typeof PathSegmentType[keyof typeof PathSegmentType];


export const PathSegmentType = {
  file: 'file',
  folder: 'folder',
  root: 'root',
} as const;

export interface PathSegment {
  /** @nullable */
  id: number | null;
  name: string;
  type: PathSegmentType;
}

export interface StorageStats {
  totalFiles: number;
  totalFolders: number;
  totalBytes: number;
  recentUploads?: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListItemsParams = {
/**
 * Parent folder ID. Omit for root.
 * @nullable
 */
parentId?: number | null;
/**
 * Search by name (full-tree search)
 */
search?: string;
};

export type ListRecentItemsParams = {
limit?: number;
};