# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Inferno cloud storage API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Items
    description: File and folder management
  - name: Storage
    description: Object storage upload and serving endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /items:
    get:
      operationId: listItems
      tags: [Items]
      summary: List items in a folder
      parameters:
        - name: parentId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
          description: Parent folder ID. Omit for root.
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name (full-tree search)
      responses:
        "200":
          description: Items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [Items]
      summary: Create a file or folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /items/recent:
    get:
      operationId: listRecentItems
      tags: [Items]
      summary: List recently uploaded or modified files
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"

  /items/{id}:
    get:
      operationId: getItem
      tags: [Items]
      summary: Get a single item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateItem
      tags: [Items]
      summary: Rename or move an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteItem
      tags: [Items]
      summary: Delete an item (recursive for folders)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /items/{id}/path:
    get:
      operationId: getItemPath
      tags: [Items]
      summary: Get full ancestor path of an item (breadcrumb)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Path segments from root to this item
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PathSegment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/stats:
    get:
      operationId: getStorageStats
      tags: [Storage]
      summary: Get storage usage statistics
      responses:
        "200":
          description: Storage stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageStats"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Item:
      type: object
      required: [id, name, type, parentId, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [file, folder]
        parentId:
          type: ["integer", "null"]
        objectPath:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ItemInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          enum: [file, folder]
        parentId:
          type: ["integer", "null"]
        objectPath:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        size:
          type: ["integer", "null"]

    ItemUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: ["integer", "null"]

    PathSegment:
      type: object
      required: [id, name, type]
      properties:
        id:
          type: ["integer", "null"]
        name:
          type: string
        type:
          type: string
          enum: [file, folder, root]

    StorageStats:
      type: object
      required: [totalFiles, totalFolders, totalBytes]
      properties:
        totalFiles:
          type: integer
        totalFolders:
          type: integer
        totalBytes:
          type: integer
        recentUploads:
          type: integer

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string