# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Inferno cloud storage API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemType } from './itemType';

export interface Item {
  id: number;
  name: string;
  type: ItemType;
  /** @nullable */
  parentId: number | null;
  /** @nullable */
  objectPath?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
  createdAt: Date;
  updatedAt: Date;
}