# lib/api-zod/src/generated/types/itemInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Inferno cloud storage API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemInputType } from './itemInputType';

export interface ItemInput {
  /** @minLength 1 */
  name: string;
  type: ItemInputType;
  /** @nullable */
  parentId?: number | null;
  /** @nullable */
  objectPath?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
}