# lib/api-zod/src/generated/types/itemUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Inferno cloud storage API
 * OpenAPI spec version: 0.1.0
 */

export interface ItemUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  parentId?: number | null;
}