# lib/api-zod/src/generated/types/pathSegment.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Inferno cloud storage API
 * OpenAPI spec version: 0.1.0
 */
import type { PathSegmentType } from './pathSegmentType';

export interface PathSegment {
  /** @nullable */
  id: number | null;
  name: string;
  type: PathSegmentType;
}