# lib/api-zod/src/generated/types/storageStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Inferno cloud storage API
 * OpenAPI spec version: 0.1.0
 */

export interface StorageStats {
  totalFiles: number;
  totalFolders: number;
  totalBytes: number;
  recentUploads?: number;
}