# lib/api-zod/src/generated/types/uploadUrlResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Inferno cloud storage API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadUrlRequest } from './uploadUrlRequest';

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}